pyproject.toml
[project]
name = "dl-lightning-template"
version = "0.1.0"
description = ""
license = "MIT"
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
   "hydra-colorlog>=1.2.0",
   "hydra-core>=1.3.2",
   "hydra-optuna-sweeper>=1.2.0",
   "lightning==2.5.0.post0",
   "rich>=13.9.4",
   "rootutils>=1.0.7",
   "sqlalchemy==1.4.44", # Fixed version to avoid issue. (cf. https://github.com/optuna/optuna/issues/4392)
   "torchvision>=0.20.1",
   "wandb>=0.19.1",
]

[dependency-groups]
tools = [
   "mypy>=1.14.0",
   "optuna-dashboard>=0.14.0",
   "bottle==0.12.25", # Fixed version to avoid issue. (cf. https://github.com/optuna/optuna-dashboard/issues/956)
   "ruff>=0.8.4",
   "taskipy>=1.14.1",
   "pre-commit>=4.0.1",
   "pytest>=8.3.4",
]

[tool.uv]
default-groups = ["tools"]
required-version = ">=0.5.18"

[tool.taskipy.tasks]
setup = "sh scripts/setup.sh"
clean-logs = "rm -rf logs/**"
test = "uv run pytest"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
   "A",
   "ANN002",
   "ANN003",
   "ARG",
   "C90",
   "D1",
   "D203",
   "D213",
   "D400",
   "EM",
   "ERA",
   "G",
   "N812",
   "PLR0913",
   "PTH",
   "RET504",
   "S101",
   "T201",
   "TRY003",
]

fixable = ["ALL"]
unfixable = ["F841"]

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["INP"]

[tool.mypy]
explicit_package_bases = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"
pythonpath = "src"
testpaths = ["tests",]
markers = [
  "slow: slow tests",
]
